Correct use_saved_calibration parameter (#226)
OLD_README.md currently refers to the parameter `use_previous_calibration`, but the correct parameter name is `use_saved_calibration`. This PR changes the existing parameter reference to the correct word, and adds the parameter to the default parameter list. 

There have been multiple Discord questions around this, so even though this PR will be outdated soon due to the Alpha release, it may save some time while pre-alpha is still in use.